+++ packages/nestjs-config-aws/src/integration/index.ts
@@ -1,6 +1,0 @@
-// Integration module exports
-export * from './interfaces';
-export * from './nestjs-config-integration.module';
-export * from './providers';
-export * from './services';
-export * from './utils';