+++ packages/nestjs-config-aws/src/integration/utils/index.ts
@@ -1,2 +1,0 @@
-// Integration utilities exports
-export * from './config-integration.util';